% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_nse.R
\name{filter_dotdata}
\alias{filter_dotdata}
\title{Using .data in a function, existing column}
\usage{
filter_dotdata(df)
}
\arguments{
\item{df}{A tibble}
}
\value{
A tibble
}
\description{
Using .data in a function, existing column
}
