% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_nse.R
\name{filter_nse}
\alias{filter_nse}
\title{Using non-standard evaluation in a function
Results in NOTE: no visible binding for global variable "state"}
\usage{
filter_nse(df)
}
\arguments{
\item{df}{A tibble}
}
\value{
A tibble
}
\description{
Using non-standard evaluation in a function
Results in NOTE: no visible binding for global variable "state"
}
